% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{centipede_data}
\alias{centipede_data}
\title{Count readstarts at each nucleotide position in each BED region.}
\usage{
centipede_data(bam_file, fimo_file, pvalue = 1e-04, flank_size = 100, ...)
}
\arguments{
\item{bam_file}{A BAM file with mapped DNase-Seq data.}

\item{fimo_file}{A FIMO text file with PWM match sites.}

\item{pvalue}{Select FIMO matches with p-value less than this.}
}
\value{
A list with two items "mat" and "regions".
}
\description{
Count readstarts at each nucleotide position in each BED region.
}
\details{
The returend matrix "mat" has one row for each region in the BED
   file, and one column for each genomic position.
   The returned dataframe "regions" described each region with a p-value
   and q-value from FIMO.
}
