% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_plots.r
\name{fitAndPlot}
\alias{fitAndPlot}
\title{Fit a motif with CENTIPEDE and plot the profile and posterior distribution.}
\usage{
fitAndPlot(motif, cen, score = NA, pvalue = NA)
}
\arguments{
\item{motif}{A string of a specific motif}

\item{cen}{an object}

\item{score}{fimo score cutoff}

\item{p.value}{p.value cutoff}
}
\value{
plot of profile and posterior distribution of a specific motif
}
\description{
Fit a motif with CENTIPEDE and plot the profile and posterior distribution.
}
\details{
cen is a list returned by \code{\link{centipede_data}}. It has a dataframe
   called \code{regions} with one row for each putative motif match.
   It also has a matrix called \code{mat} with one row for each match,
   and each column is a nucleotide position relative to the center of the
   motif match. The values in this matrix are the count of read start sites
   at that position.
}

