% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_bedGraph}
\alias{read_bedGraph}
\title{Read a bedGraph file with 4 columns: chrom, start, end, score}
\usage{
read_bedGraph(filename, ...)
}
\arguments{
\item{filename}{The file to read.}

\item{...}{Additional parameters passed to read.delim.}
}
\value{
A GRanges object.
}
\description{
Read a bedGraph file with 4 columns: chrom, start, end, score
}

