% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_fimo}
\alias{read_fimo}
\title{Read a text file output by FIMO and select sites that meet a significance
threshold.}
\usage{
read_fimo(fimo_file, pvalue = 1e-04, ...)
}
\arguments{
\item{fimo_file}{A FIMO text file with PWM match sites.}

\item{pvalue}{Select FIMO matches with p-value less than this.}
}
\value{
A dataframe with one row per site.
}
\description{
Read a text file output by FIMO and select sites that meet a significance
threshold.
}
\details{
For sites represented multiple times, the one with the maximum
   score is selected.
}

