% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_cons}
\alias{site_cons}
\title{A vector of mean conservation scores for the PWM sites in "cen"}
\format{A vector that corresponds to the data in "cen". Each item in the
vector is the mean conservation score across 100 verterbrates computed
with phastCons at UCSC.}
\source{
\url{http://hgdownload.cse.ucsc.edu/goldenpath/hg19/phastCons100way/}
}
\usage{
site_cons
}
\description{
A vector of mean conservation scores for the PWM sites in "cen"
}
\keyword{datasets}

