% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_bedGraph}
\alias{read_bedGraph}
\title{Read a bedGraph file with 4 columns: chrom, start, end, score}
\usage{
read_bedGraph(filename, ...)
}
\arguments{
\item{filename}{The file to read. No column names.}

\item{...}{Additional parameters passed to \code{\link[readr]{read_tsv}}}
}
\value{
A GRanges object.
}
\description{
Read a bedGraph file with 4 columns: chrom, start, end, score
}
