% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_fimo}
\alias{read_fimo}
\title{Read a text file output by FIMO and select sites that meet a significance
threshold.}
\usage{
read_fimo(fimo_file, log10p = 4)
}
\arguments{
\item{fimo_file}{A FIMO text file with PWM match sites.}

\item{log10p}{Select FIMO matches with log10 p-value greater than this.}
}
\value{
A dataframe with one row per site.
}
\description{
Read a text file output by FIMO and select sites that meet a significance
threshold.
}
\details{
For sites represented multiple times, the one with the maximum
   score is selected.
}

